requirements.txt
streamlit
pandas
numpy
matplotlib
openpyxl
requests
Pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
google-api-python-client
google-cloud-storage
oauth2client
google-api-python-client